refactored operator selection to memoize properly and added null check to filter building
